chore: Refactor Parse Command
