+++ internal/commands/parse.go
@@ -1,12 +1,10 @@
 package commands
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"strings"
 
+	"github.com/open-policy-agent/conftest/internal/runner"
 	"github.com/open-policy-agent/conftest/parser"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -43,7 +41,9 @@ func NewParseCommand(ctx context.Context) *cobra.Command {
 			return nil
 		},
 		RunE: func(cmd *cobra.Command, fileList []string) error {
-			out, err := parseInput(ctx, viper.GetString("input"), viper.GetBool("combine"), fileList)
+			runner := &runner.ParseRunner{}
+			viper.Unmarshal(runner)
+			out, err := runner.Run(ctx, fileList)
 			if err != nil {
 				return fmt.Errorf("failed during parser process: %w", err)
 			}
@@ -56,60 +56,4 @@ func NewParseCommand(ctx context.Context) *cobra.Command {
 	cmd.Flags().BoolP("combine", "", false, "combine all given config files to be evaluated together")
 	cmd.Flags().StringP("input", "i", "", fmt.Sprintf("input type for given source, especially useful when using conftest with stdin, valid options are: %s", parser.ValidInputs()))
 	return &cmd
-}
-
-func parseInput(ctx context.Context, input string, combine bool, fileList []string) (string, error) {
-	configurations, err := parser.GetConfigurations(ctx, input, fileList)
-	if err != nil {
-		return "", fmt.Errorf("calling the parser method: %w", err)
-	}
-
-	parsedConfigurations, err := parseConfigurations(configurations, combine)
-	if err != nil {
-		return "", fmt.Errorf("parsing configs: %w", err)
-	}
-
-	return parsedConfigurations, nil
-
-}
-
-func parseConfigurations(configurations map[string]interface{}, combine bool) (string, error) {
-	var output string
-	if combine {
-		content, err := marshal(configurations)
-		if err != nil {
-			return "", fmt.Errorf("marshal output to json: %w", err)
-		}
-
-		output = strings.Replace(output+"\n"+content, "\\r", "", -1)
-	} else {
-		for filename, config := range configurations {
-			content, err := marshal(config)
-			if err != nil {
-				return "", fmt.Errorf("marshal output to json: %w", err)
-			}
-
-			output = strings.Replace(output+filename+"\n"+content, "\\r", "", -1)
-		}
-	}
-
-	return output, nil
-}
-
-func marshal(in interface{}) (string, error) {
-	out, err := json.Marshal(in)
-	if err != nil {
-		return "", fmt.Errorf("marshal output to json: %w", err)
-	}
-
-	var prettyJSON bytes.Buffer
-	if err = json.Indent(&prettyJSON, out, "", "\t"); err != nil {
-		return "", fmt.Errorf("indentation: %w", err)
-	}
-
-	if _, err := prettyJSON.WriteString("\n"); err != nil {
-		return "", fmt.Errorf("adding line break: %w", err)
-	}
-
-	return prettyJSON.String(), nil
 }
