+++ internal/runner/parse_test.go
@@ -1,4 +1,4 @@
-package commands
+package runner
 
 import (
 	"strings"
